# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "企业知识库问答系统 API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "登录用户的 MongoDB _id"
      },
      "UserName": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-name",
        "description": "登录用户名"
      },
      "UserRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "用户角色：admin 或 user"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "获取用户列表",
        "description": "仅管理员可访问，支持 keyword、role、status 筛选",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            },
            "description": "按用户名或昵称模糊搜索"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "disabled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "新增用户",
        "description": "仅管理员可操作，默认密码为 123456",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "default": "123456"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "user"
                    ],
                    "default": "user"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "disabled"
                    ],
                    "default": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "参数错误或用户名已存在"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "获取用户详情",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "403": {
            "description": "无权限"
          },
          "404": {
            "description": "用户不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "更新用户",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "user"
                    ]
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "disabled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "参数错误"
          },
          "403": {
            "description": "无权限"
          },
          "404": {
            "description": "用户不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "删除用户",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "403": {
            "description": "无权限"
          },
          "404": {
            "description": "用户不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/users/{id}/reset-password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "重置用户密码",
        "description": "仅管理员可操作，重置为默认密码 123456",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "密码重置成功"
          },
          "403": {
            "description": "无权限"
          },
          "404": {
            "description": "用户不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "切换用户状态",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "disabled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "状态更新成功"
          },
          "403": {
            "description": "无权限"
          },
          "404": {
            "description": "用户不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  }
}